PreprocessMD

Documentation for PreprocessMD.jl.

PreprocessMD

Example usage

Function index

Function documentation

Example usage

using CSV: File
using Downloads: download

using DataFrames, MLJ, PreprocessMD

# Download synthetic data
url = "https://physionet.org/files/mimic-iv-demo-omop/0.9/1_omop_data_csv"

# Read in files as DataFrames
DRUG = DataFrame(File.(download("$url/drug_exposure.csv")));
CONDITION = DataFrame(File.(download("$url/condition_occurrence.csv")));
DEATH = DataFrame(File.(download("$url/death.csv")));

# Convert to wide format
p_CONDITION = pivot(CONDITION, :person_id, :condition_concept_id);
p_DRUG = pivot(DRUG, :person_id, :drug_concept_id);

# Join tables by patient
p_AGGREGATE = innerjoin(p_CONDITION, p_DRUG, on=:person_id);

# Add label data to feature set
add_label_column!(p_AGGREGATE, DEATH, :death)

# ... Continue to machine learning pipeline ...

Function index

PreprocessMD.PreprocessMD

PreprocessMD.generate_cohort

PreprocessMD.pivot

PreprocessMD.subsetMD

PreprocessMD.subset_invalid_year

PreprocessMD.top_n_values

Function documentation

PreprocessMD.PreprocessMD — Module

Description

Medically-informed data preprocessing for machine learning

PreprocessMD.generate_cohort — Method

Function -> Generatecohort Inputs Colname -> column name domain_table -> the case table on which your concepts are based on example conditions, observations, drugs, measurements concepts -> The list of concept id's you want to study Output Unique list of person ids in your cohort whom you want to study

Examples

julia> using DataFrames

julia> df_condition_occurrence = DataFrame(condition_occurrence_id=[123, 5433, 8765, 12345, 6457, 62898], person_id = [1, 2, 3, 4, 5, 6], condition_concept_id = [196523, 436659, 435515, 436096, 440383, 37311319])
6×3 DataFrame
 Row │ condition_occurrence_id  person_id  condition_concept_id
     │ Int64                    Int64      Int64
─────┼──────────────────────────────────────────────────────────
   1 │                     123          1                196523
   2 │                    5433          2                436659
   3 │                    8765          3                435515
   4 │                   12345          4                436096
   5 │                    6457          5                440383
   6 │                   62898          6              37311319

julia> concepts = [196523, 436659, 435515, 436096, 440383]
5-element Vector{Int64}:
 196523
 436659
 435515
 436096
 440383

julia> result = generate_cohort( :condition_concept_id, df_condition_occurrence, concepts)
5-element Vector{Int64}:
 1
 2
 3
 4
 5

PreprocessMD.pivot — Function

function pivot()

Express the long format DataFrame df as a wide format DataFrame B.

Optional arguments x and y are columns of df. The single column x (the first column of df, by default) becomes the row names of B. Column(s) y (all columns besides x, by default) become the column names of B.

Examples

julia> using DataFrames

julia> df = DataFrame(name=["Cookie Monster", "Elmo", "Oscar", "Grover"], fur_color=["blue", "red", "green", "blue"])
4×2 DataFrame
 Row │ name            fur_color
     │ String          String
─────┼───────────────────────────
   1 │ Cookie Monster  blue
   2 │ Elmo            red
   3 │ Oscar           green
   4 │ Grover          blue

julia> pivot(df)
4×4 DataFrame
 Row │ name            blue   red    green
     │ String          Bool   Bool   Bool
─────┼─────────────────────────────────────
   1 │ Cookie Monster   true  false  false
   2 │ Elmo            false   true  false
   3 │ Oscar           false  false   true
   4 │ Grover           true  false  false

PreprocessMD.subsetMD — Function

function subsetMD(main_df, check_df, main_id, check_id)

Filtration step

Arguments

main_df::AbstractDataFrame: Rows are selected from this DataFrame...

check_df::AbstractDataFrame: ... if the IDs are present in this DataFrame

main_id: ID column from main_df (Default: first column)

check_id: ID column from check_df (Default: same as main_id)

Examples

julia> using DataFrames

julia> X = DataFrame(name=["Cookie Monster", "Elmo", "Oscar", "Grover"], blue = [true, false, false, true], red  = [false, true, false, false], green = [false, false, true, false])
4×4 DataFrame
 Row │ name            blue   red    green
     │ String          Bool   Bool   Bool
─────┼─────────────────────────────────────
   1 │ Cookie Monster   true  false  false
   2 │ Elmo            false   true  false
   3 │ Oscar           false  false   true
   4 │ Grover           true  false  false

julia> Y = DataFrame(name=["Big Bird", "Cookie Monster", "Elmo"], fuzzy=[false, true, true])
3×2 DataFrame
 Row │ name            fuzzy
     │ String          Bool
─────┼───────────────────────
   1 │ Big Bird        false
   2 │ Cookie Monster   true
   3 │ Elmo             true

julia> subsetMD(X,Y)
2×4 DataFrame
 Row │ name            blue   red    green
     │ String          Bool   Bool   Bool
─────┼─────────────────────────────────────
   1 │ Cookie Monster   true  false  false
   2 │ Elmo            false   true  false

PreprocessMD.subset_invalid_year — Method

function subset_invalid_year

Arguments

Examples

julia> using PreprocessMD, Dates, DataFrames

julia> X = DataFrame(name=["Kermit", "Big Bird", "Herry", "Mr. Snuffleupagus", "Rosita", "Julia"], first_appearance=Date.(["1955-05-09", "1969-11-10", "1970-11-09", "1971-11-15", "1991-11-26", "2017-04-10"]))
6×2 DataFrame
 Row │ name               first_appearance
     │ String             Dates.Date
─────┼─────────────────────────────────────
   1 │ Kermit             1955-05-09
   2 │ Big Bird           1969-11-10
   3 │ Herry              1970-11-09
   4 │ Mr. Snuffleupagus  1971-11-15
   5 │ Rosita             1991-11-26
   6 │ Julia              2017-04-10

julia> subset_invalid_year(X, :first_appearance, 1969, 2000)
2×2 DataFrame
 Row │ name    first_appearance
     │ String  Dates.Date
─────┼──────────────────────────
   1 │ Kermit  1955-05-09
   2 │ Julia   2017-04-10

PreprocessMD.top_n_values — Function

function top_n_values(df::AbstractDataFrame, col::Union{String, Symbol}, n::Int)::AbstractDataFrame

Find top n values by occurence Useful for initial feasibility checks, but medical codes are not considered

Examples

julia> using DataFrames

julia> df = DataFrame(name=["Cookie Monster", "Elmo", "Oscar", "Grover", "Big Bird", "Ernie", "Bert", "Rosita"], fur_color=["blue", "red", "green", "blue", "yellow", "orange", "yellow", "blue"])
8×2 DataFrame
 Row │ name            fur_color
     │ String          String
─────┼───────────────────────────
   1 │ Cookie Monster  blue
   2 │ Elmo            red
   3 │ Oscar           green
   4 │ Grover          blue
   5 │ Big Bird        yellow
   6 │ Ernie           orange
   7 │ Bert            yellow
   8 │ Rosita          blue

julia> top_n_values(df, :fur_color, 4)
4×2 DataFrame
 Row │ fur_color  nrow
     │ String     Int64
─────┼──────────────────
   1 │ blue           3
   2 │ yellow         2
   3 │ red            1
   4 │ green          1