PreprocessMD

Documentation for PreprocessMD.jl.

PreprocessMD

Example usage

Function index

Function documentation

Example usage

using CSV: File
using DataFrames: innerjoin
using DataFrames: DataFrame
using Downloads: download

using PreprocessMD: add_label_column!
using PreprocessMD: pivot

# Download synthetic data
url = "https://physionet.org/files/mimic-iv-demo-omop/0.9/1_omop_data_csv"

# Read in files as DataFrames
DRUG = DataFrame(File.(download("$url/drug_exposure.csv")));
CONDITION = DataFrame(File.(download("$url/condition_occurrence.csv")));
DEATH = DataFrame(File.(download("$url/death.csv")));

# Convert to wide format
p_CONDITION = pivot(CONDITION, :person_id, :condition_concept_id);
p_DRUG = pivot(DRUG, :person_id, :drug_concept_id);

# Join tables by patient
p_AGGREGATE = innerjoin(p_CONDITION, p_DRUG, on=:person_id);

# Add label data to feature set
add_label_column!(p_AGGREGATE, DEATH, :death)

# ... Continue to machine learning pipeline ...

Function index

PreprocessMD.PreprocessMD

PreprocessMD.add_label_column!

PreprocessMD.pivot

PreprocessMD.set_label_column!

PreprocessMD.subsetMD

PreprocessMD.top_n_values

Function documentation

PreprocessMD.PreprocessMD — Module

Description

Medically-informed data preprocessing for machine learning

Structure

PreprocessMD:

add_label_column!

pivot

set_label_column!

subsetMD

top_n_values

PreprocessMD.add_label_column! — Function

function add_label_column!(feature_df, source_df, new_column[, id])

Add a label column to a DataFrame based on symbol presence in the target DataFrame

A column from the target is not copied. Instead, the new column is a CategoricalArray containing true for any ID that is present in the target and false otherwise.

Arguments

feature_df::AbstractDataFrame: feature DataFrame to which label column is added

source_df::AbstractDataFrame: DataFrame containing the label column

new_column::Union{String, Symbol}: name assigned to label column

id::Union{Nothing, String, Symbol}: row IDs (Default: first column)

Examples

julia> using DataFrames

julia> X = DataFrame(name=["Cookie Monster", "Elmo", "Oscar", "Grover"], blue = [true, false, false, true], red  = [false, true, false, false], green = [false, false, true, false])
4×4 DataFrame
 Row │ name            blue   red    green
     │ String          Bool   Bool   Bool
─────┼─────────────────────────────────────
   1 │ Cookie Monster   true  false  false
   2 │ Elmo            false   true  false
   3 │ Oscar           false  false   true
   4 │ Grover           true  false  false

julia> Y = DataFrame(name=["Cookie Monster", "Elmo", "Oscar", "Grover"], lovable = [true, true, false, true], furry = [true, true, true, true], old = [false, false, true, true])
4×4 DataFrame
 Row │ name            lovable  furry  old
     │ String          Bool     Bool   Bool
─────┼───────────────────────────────────────
   1 │ Cookie Monster     true   true  false
   2 │ Elmo               true   true  false
   3 │ Oscar             false   true   true
   4 │ Grover             true   true   true

julia> add_label_column!(X,Y,:furry)

julia> X
4×5 DataFrame
 Row │ name            blue   red    green  furry
     │ String          Bool   Bool   Bool   Cat…
─────┼────────────────────────────────────────────
   1 │ Cookie Monster   true  false  false  true
   2 │ Elmo            false   true  false  true
   3 │ Oscar           false  false   true  true
   4 │ Grover           true  false  false  true

PreprocessMD.pivot — Function

function pivot()

Express the long format DataFrame df as a wide format DataFrame B.

Optional arguments x and y are columns of df. The single column x (the first column of df, by default) becomes the row names of B. Column(s) y (all columns besides x, by default) become the column names of B.

Examples

julia> using DataFrames

julia> df = DataFrame(name=["Cookie Monster", "Elmo", "Oscar", "Grover"], fur_color=["blue", "red", "green", "blue"])
4×2 DataFrame
 Row │ name            fur_color
     │ String          String
─────┼───────────────────────────
   1 │ Cookie Monster  blue
   2 │ Elmo            red
   3 │ Oscar           green
   4 │ Grover          blue

julia> pivot(df)
4×4 DataFrame
 Row │ name            blue   red    green
     │ String          Bool   Bool   Bool
─────┼─────────────────────────────────────
   1 │ Cookie Monster   true  false  false
   2 │ Elmo            false   true  false
   3 │ Oscar           false  false   true
   4 │ Grover           true  false  false

PreprocessMD.set_label_column! — Function

function set_label_column!(feature_df, source_df, new_column[, id])

Designate one column within a DataFrame as the label

Arguments

feature_df::AbstractDataFrame: feature DataFrame

col_name::Union{String, Symbol}: label column

id::Union{Nothing, String, Symbol}: row IDs (Default: first column)

Examples

julia> using DataFrames

julia> X = DataFrame(name=["Cookie Monster", "Elmo", "Oscar", "Grover"], lovable = [true, true, false, true], furry = [true, true, true, true], old = [false, false, true, true] )
4×4 DataFrame
 Row │ name            lovable  furry  old
     │ String          Bool     Bool   Bool
─────┼───────────────────────────────────────
   1 │ Cookie Monster     true   true  false
   2 │ Elmo               true   true  false
   3 │ Oscar             false   true   true
   4 │ Grover             true   true   true

julia> set_label_column!(X,:lovable)

julia> X
4×4 DataFrame
 Row │ name            lovable  furry  old
     │ String          Cat…     Bool   Bool
─────┼───────────────────────────────────────
   1 │ Cookie Monster  true      true  false
   2 │ Elmo            true      true  false
   3 │ Oscar           false     true   true
   4 │ Grover          true      true   true

PreprocessMD.subsetMD — Function

function subsetMD(main_df, check_df, main_id, check_id)

Filtration step

Arguments

main_df::AbstractDataFrame: Rows are selected from this DataFrame...

check_df::AbstractDataFrame: ... if the IDs are present in this DataFrame

main_id: ID column from main_df (Default: first column)

check_id: ID column from check_df (Default: same as main_id)

Examples

julia> using DataFrames

julia> X = DataFrame(name=["Cookie Monster", "Elmo", "Oscar", "Grover"], blue = [true, false, false, true], red  = [false, true, false, false], green = [false, false, true, false])
4×4 DataFrame
 Row │ name            blue   red    green
     │ String          Bool   Bool   Bool
─────┼─────────────────────────────────────
   1 │ Cookie Monster   true  false  false
   2 │ Elmo            false   true  false
   3 │ Oscar           false  false   true
   4 │ Grover           true  false  false

julia> Y = DataFrame(name=["Big Bird", "Cookie Monster", "Elmo"], fuzzy=[false, true, true])
3×2 DataFrame
 Row │ name            fuzzy
     │ String          Bool
─────┼───────────────────────
   1 │ Big Bird        false
   2 │ Cookie Monster   true
   3 │ Elmo             true

julia> subsetMD(X,Y)
2×4 DataFrame
 Row │ name            blue   red    green
     │ String          Bool   Bool   Bool
─────┼─────────────────────────────────────
   1 │ Cookie Monster   true  false  false
   2 │ Elmo            false   true  false

PreprocessMD.top_n_values — Function

function top_n_values(df::AbstractDataFrame, col::Union{String, Symbol}, n::Int)::AbstractDataFrame

Find top n values by occurence Useful for initial feasibility checks, but medical codes are not considered

Examples

julia> using DataFrames

julia> df = DataFrame(name=["Cookie Monster", "Elmo", "Oscar", "Grover", "Big Bird", "Ernie", "Bert", "Rosita"], fur_color=["blue", "red", "green", "blue", "yellow", "orange", "yellow", "blue"])
8×2 DataFrame
 Row │ name            fur_color
     │ String          String
─────┼───────────────────────────
   1 │ Cookie Monster  blue
   2 │ Elmo            red
   3 │ Oscar           green
   4 │ Grover          blue
   5 │ Big Bird        yellow
   6 │ Ernie           orange
   7 │ Bert            yellow
   8 │ Rosita          blue

julia> top_n_values(df, :fur_color, 4)
4×2 DataFrame
 Row │ fur_color  nrow
     │ String     Int64
─────┼──────────────────
   1 │ blue           3
   2 │ yellow         2
   3 │ red            1
   4 │ green          1