PreprocessMD

Documentation for PreprocessMD.jl.

PreprocessMD

Example usage

Function index

Function documentation

Example usage

using CSV: File
using DataFrames: innerjoin
using DataFrames: DataFrame
using Downloads: download

using PreprocessMD: add_label_column!
using PreprocessMD: MLDemo
using PreprocessMD: pivot

url = "https://physionet.org/files/mimic-iv-demo-omop/0.9/1_omop_data_csv"
DRUG = DataFrame(File.(download("$url/drug_exposure.csv")));
CONDITION = DataFrame(File.(download("$url/condition_occurrence.csv")));
DEATH = DataFrame(File.(download("$url/death.csv")));

p_CONDITION = pivot(CONDITION, :person_id, :condition_concept_id);
p_DRUG = pivot(DRUG, :person_id, :drug_concept_id);

p_AGGREGATE = innerjoin(p_CONDITION, p_DRUG, on=:person_id);

add_label_column!(p_AGGREGATE, DEATH, :death)

MLDemo(p_AGGREGATE, :death, 1234) |> display

Evaluating over 6 folds:  33%[========>                ]  ETA: 0:01:31Evaluating over 6 folds: 100%[=========================] Time: 0:00:46
[ Info: Training machine(DecisionTreeClassifier(max_depth = 3, …), …).

Function index

PreprocessMD.MLDemo

PreprocessMD.add_label_column!

PreprocessMD.pivot

PreprocessMD.subsetMD

PreprocessMD.top_n_values

Function documentation

PreprocessMD.MLDemo — Method

function MLDemo(df::AbstractDataFrame, output::Symbol, RNG_VALUE)::Tuple{AbstractFloat, AbstractFloat}

Decision tree classifier on a DataFrame over a given output

Arguments

df::AbstractDataFrame: DataFrame containing feature and label data

output::Symbol: column containing label data

RNG_VALUE:

PreprocessMD.add_label_column! — Function

function add_label_column!(to_df, from_df, new_col_name[, id])::Nothing

Add column to a DataFrame based on symbol presence in the target DataFrame

Arguments

`

to_df::AbstractDataFrame: feature DataFrame to which label column is added

from_df::AbstractDataFrame: DataFrame containing the label column

new_col_name: name assigned to label column

id=nothing: row IDs (Default: first column)

Examples

X = DataFrame(name=["Cookie Monster", "Elmo", "Oscar", "Grover"],
	blue = [true, false, false, true],
	red  = [false, true, false, false],
	green = [false, false, true, false]);

Y = DataFrame(name=["Cookie Monster", "Elmo", "Oscar", "Grover"],
	lovable = [true, true, false, true],
	furry = [true, true, true, true],
	old = [false, false, true, true]
	);
add_label_column!(X,Y,:furry)
X

# output
4×5 DataFrame
 Row │ name            blue   red    green  furry 
     │ String          Bool   Bool   Bool   Cat…  
─────┼────────────────────────────────────────────
   1 │ Cookie Monster   true  false  false  true
   2 │ Elmo            false   true  false  true
   3 │ Oscar           false  false   true  true
   4 │ Grover           true  false  false  true

PreprocessMD.pivot — Function

function pivot()

Express the long format DataFrame df as a wide format DataFrame B.

Optional arguments x and y are columns of df. The single column x (the first column of df, by default) becomes the row names of B. Column(s) y (all columns besides x, by default) become the column names of B.

Examples

df = DataFrame(name=["Cookie Monster", "Elmo", "Oscar", "Grover"], 
               fur_color=["blue", "red", "green", "blue"]);
pivot(df)

# output
4×4 DataFrame
 Row │ name            blue   red    green 
     │ String          Bool   Bool   Bool  
─────┼─────────────────────────────────────
   1 │ Cookie Monster   true  false  false
   2 │ Elmo            false   true  false
   3 │ Oscar           false  false   true
   4 │ Grover           true  false  false

PreprocessMD.subsetMD — Function

function subsetMD(main_df, check_df, main_id, check_id)

Filtration step

Arguments

main_df::AbstractDataFrame: Rows are selected from this DataFrame...

check_df::AbstractDataFrame: ... if the IDs are present in this DataFrame

main_id: ID column from main_df (Default: first column)

check_id: ID column from check_df (Default: same as main_id)

Examples

X = DataFrame(
	name=["Cookie Monster", "Elmo", "Oscar", "Grover"],
	blue = [true, false, false, true],
	red  = [false, true, false, false],
	green = [false, false, true, false]);

Y = DataFrame(
	name=["Big Bird", "Cookie Monster", "Elmo"],
	fuzzy=[false, true, true]
	);
subsetMD(X,Y)

# output
2×4 DataFrame
 Row │ name            blue   red    green 
     │ String          Bool   Bool   Bool  
─────┼─────────────────────────────────────
   1 │ Cookie Monster   true  false  false
   2 │ Elmo            false   true  false

PreprocessMD.top_n_values — Method

function top_n_values(df::AbstractDataFrame, col::Symbol, n::Int)::AbstractDataFrame

Find top n values by occurence Useful for initial feasibility checks, but medical codes are not considered